Disabled writing of FreeMarker logs to stdout when generating an HTML report
Due to an untraceable error of FreeMarker templates, repeated attempts to generate a report are triggered in some cases. In case of unsuccessful attempts, FreeMarker writes to its logger, which is sent to stdout by default.
If the error repeats frequently, an exception will be thrown with an error message - so we can ignore FreeMarker's own logging when generating an HTML report.

Fixes #446