Fixed applying of the root verification rules
Now, when creating the Android report configuration, on Check is immediately set for verification.

This is not necessary, because if not specified, only the `koverVerify` task is started when executing check, but not the tasks of verifying Android build variant.
This has a negative side effect, because the verification for the build variant is always specify, and it will override the common rules from the root of the report config `koverReport { verify { /* common rules */ }  }`

Fixes #459