Fixed `Stream closed` error  when generating Kover HTML report
When generating an HTML report, errors sometimes occur when reading the freemarker template from resources. The exact cause of the error has not yet been determined (relates to the Gradle isolated class loaders), but it disappears when you try to read again.
In some cases, it is necessary to increase the number of repetitions, so this setting was made in Kover Gradle plugin.

Fixes #510