Repacked the intellij-agent artifact to kover-jvm-agent
User can use the JVM agent separately, with the name kover and without mentioning IntelliJ, as well as reduce the number of parameters and properties required to transfer to the agent.

Now Kover Gradle Plugin uses kover-jvm-agent with the same version. To do this, the transfer of staging repositories was added to the functional tests.
The common logic of publishing was also extracted into a separate convention plugin.

Resolves #464