Fixed error: Kover requires extension with name 'androidComponents'
In older versions of AGP (lower than 7.0), the `androidComponents` project extension did not exist and this is the expected situation. However, in the current implementation, in the absence of this extension, the KoverCriticalException error was thrown.

In this case, it is necessary not to throw an error, but execute fallback branch

Fixes #520