Used compile dependency to Kover Features in Kover Gradle Plugin
Since the version of the aggregator and report generator has been enshrined in the Kover Gradle Plugin for a long time, it makes no sense to use dynamic classpath of and classloader isolations in Gradle Workers.

We can use Kover Features as a compile dependency and use its API.

This change can also eliminate the error with access to Freemarker template inside Gradle Worker, and simplify the implementation of warn on verification error.

Resolves #567