Expose reportDir property in KoverHtmlReport interface
if there is a need to analyze the HTML report after it is generated, or if you need to improve the report (for example, by editing CSS styles), you need to know the directory of each report.
Currently, it is not possible to easily link a specific report generation task and a directory.

Resolves #590